+++ apps/frontend/src/app/3_organisms/LOCChart/LOCChart.tsx
@@ -11,17 +11,11 @@ import {
   Legend,
 } from 'chart.js';
 import { Bar } from 'react-chartjs-2';
-import { useTranslation } from 'react-i18next';
 import { useLoaderData } from 'react-router-dom';
-
-import { SupportedTokens } from '@sovryn/contracts';
-import { prettyTx } from '@sovryn/ui';
 
 import { ZeroPageLoaderData } from '../../5_pages/ZeroPage/loader';
 import { useAccount } from '../../../hooks/useAccount';
-import { translations } from '../../../locales/i18n';
-import { Bitcoin } from '../../../utils/constants';
-import { formatValue } from '../../../utils/math';
+import { useGetChartOptions } from './hooks/useGetChartOptions';
 import { useGetLowestTroves } from './hooks/useGetLowestTroves';
 import { useGetRBTCPrice } from './hooks/useGetRBTCPrice';
 import { useGetTroves } from './hooks/useGetTroves';
@@ -71,145 +65,16 @@ ChartJS.register(
  */
 
 export const LOCChart: FC = () => {
-  const { t } = useTranslation();
   const { account } = useAccount();
   const [data, setData] = useState<ChartDataStructure>([]);
   const [activeBar, setActiveBar] = useState<number | null>(null);
   const { price } = useGetRBTCPrice();
   const [userCollateralRatio, setUserCollateralRatio] = useState('');
   const [redemptionBuffer, setRedemptionBuffer] = useState(0);
-  const [xLabels, setXLabels] = useState(0);
+  const [userTrovesBelowCount, setUserTrovesBelowCount] = useState(0);
   const { liquity } = useLoaderData() as ZeroPageLoaderData;
   const [userDebtAmount, setUserDebtAmount] = useState('0');
   const [userCollateralAmount, setUserCollateralAmount] = useState('0');
-
-  const options = useMemo(() => {
-    return {
-      responsive: true,
-      maintainAspectRatio: false,
-      plugins: {
-        tooltip: {
-          borderRadius: chartConfig.tooltipBorderRadius,
-          yAlign: 'bottom' as const,
-          backgroundColor: chartConfig.fontColor,
-          padding: chartConfig.tooltipPadding,
-          caretSize: chartConfig.caretSize,
-          bodyColor: chartConfig.borderColor,
-          bodyFont: {
-            family: chartConfig.defaultFont,
-            weight: chartConfig.weight,
-            lineHeight: chartConfig.tooltipLineHeight,
-          },
-          callbacks: {
-            title: () => '',
-            label: () => '',
-            beforeBody: context => {
-              const { collateralAmount, address, debtAmount, collateralRatio } =
-                context[0].raw;
-              const tooltipContent = [
-                prettyTx(address),
-                `${t(translations.chart.collateralAmount)}: ${formatValue(
-                  Number(collateralAmount),
-                  6,
-                )} ${Bitcoin}`,
-                `${t(translations.chart.debtAmount)}: ${formatValue(
-                  Number(debtAmount),
-                  2,
-                )} ${SupportedTokens.zusd.toUpperCase()}`,
-                `${t(translations.chart.collateralRatio)}: ${formatValue(
-                  Number(collateralRatio),
-                  0,
-                )}%`,
-              ];
-
-              if (context[0].raw.sequenceNumber === userCollateralRatio) {
-                tooltipContent.push(
-                  `${t(translations.chart.redemptionBuffer)}: ${formatValue(
-                    redemptionBuffer,
-                    2,
-                  )} ${SupportedTokens.zusd.toUpperCase()}`,
-                );
-              }
-              return tooltipContent;
-            },
-          },
-        },
-        legend: {
-          labels: {
-            color: chartConfig.fontColor,
-            boxWidth: chartConfig.legendBoxWidth,
-            generateLabels: () => {
-              const labelDefault = {
-                text: t(translations.chart.collateralRatio),
-                fontColor: chartConfig.fontColor,
-                fillStyle: chartConfig.defaultColor,
-                lineWidth: chartConfig.legendLineWidth,
-              };
-              const labelActive = {
-                text: t(translations.chart.myCollateralRatio),
-                fontColor: chartConfig.fontColor,
-                fillStyle: chartConfig.activeColor,
-                lineWidth: chartConfig.legendLineWidth,
-              };
-              return !activeBar ? [labelDefault] : [labelDefault, labelActive];
-            },
-          },
-        },
-      },
-      scales: {
-        x: {
-          suggestedMax: chartConfig.maxValue,
-          suggestedMin: chartConfig.minValue,
-          grid: {
-            drawOnChartArea: false,
-            drawTicks: true,
-            tickColor: chartConfig.defaultColor,
-            tickLength: chartConfig.tickLength,
-            lineWidth: chartConfig.lineWidth,
-            offset: false,
-          },
-          border: {
-            color: chartConfig.defaultColor,
-          },
-          ticks: {
-            callback: value => Number(xLabels + value + 1).toFixed(0),
-            color: chartConfig.fontColor,
-            font: {
-              family: chartConfig.defaultFont,
-              size: chartConfig.fontSize,
-              weight: chartConfig.weight,
-            },
-            padding: chartConfig.tickPadding,
-            beginAtZero: true,
-          },
-        },
-        y: {
-          gridDashType: 'dash',
-          grid: {
-            drawOnChartArea: false,
-            drawTicks: true,
-            tickColor: chartConfig.defaultColor,
-            tickLength: chartConfig.tickLength,
-            lineWidth: chartConfig.lineWidth,
-          },
-          border: {
-            color: chartConfig.defaultColor,
-          },
-          ticks: {
-            callback: value => Number(value).toFixed(3) + '%',
-            color: chartConfig.fontColor,
-            font: {
-              family: chartConfig.defaultFont,
-              size: chartConfig.tickSize,
-              weight: chartConfig.weight,
-            },
-            padding: chartConfig.tickPadding,
-            beginAtZero: true,
-          },
-        },
-      },
-    };
-  }, [t, activeBar, userCollateralRatio, redemptionBuffer, xLabels]);
 
   const { data: userOpenTrove, loading: loadingUserOpenTrove } =
     useGetUserOpenTrove();
@@ -223,6 +88,13 @@ export const LOCChart: FC = () => {
     }
     return false;
   }, [userOpenTrove, account]);
+
+  const options = useGetChartOptions(
+    userCollateralRatio,
+    redemptionBuffer,
+    activeBar,
+    userTrovesBelowCount,
+  );
 
   useEffect(() => {
     if (!loadingLowestTroves) {
@@ -362,7 +234,9 @@ export const LOCChart: FC = () => {
       setData(sortData([...trovesDataBelow, ...userTrove, ...trovesDataAbove]));
 
       if (lowestTroves) {
-        setXLabels(lowestTroves.troves.length - trovesDataAbove.length);
+        setUserTrovesBelowCount(
+          lowestTroves.troves.length - trovesDataAbove.length,
+        );
       }
 
       setActiveBar(
@@ -400,7 +274,7 @@ export const LOCChart: FC = () => {
           price,
         ),
       }));
-      setXLabels(0);
+      setUserTrovesBelowCount(0);
       setData(sortData([...trovesData]));
     }
   }, [price, troves, loadingTroves, isUserOpenTrove, loadingUserOpenTrove]);
